use reference keys instead of relations
